BH-101 Refactor `CollisionManager` functionality into `Collider` class and update related tests
